pyproject.toml
[project]
name = "boomrdbox"
version = "0.1.0"
description = "Record and replay Redis stream data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0,<6.0",
    "msgpack>=1.0,<2.0",
    "hydra-core>=1.3,<2.0",
    "hydra-zen>=0.13",
    "omegaconf>=2.3",
    "pyarrow>=15.0",
    "tqdm>=4.66",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "pytest-xdist",
    "pytest-cov",
    "mypy",
    "ruff",
    "pre-commit",
    "vulture",
    "import-linter",
    "types-pyyaml",
    "types-redis",
    "types-tqdm",
]

[project.scripts]
boomrdbox = "boomrdbox.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/boomrdbox"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",
    "ISC001",
    "EM",
    "TRY003",
    "D213",
    "D203",
    "PLR2004",
    "PERF203",
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "D", "ANN"]
"src/boomrdbox/cli.py" = ["PLC0415"]
"src/boomrdbox/tools.py" = ["PLC0415"]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["msgpack.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pyarrow.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["hydra_zen.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -n auto --cov=boomrdbox --cov-report=term-missing"
markers = ["integration: integration tests requiring Docker Redis"]

[tool.vulture]
min_confidence = 80
paths = ["src/boomrdbox"]

[tool.importlinter]
root_package = "boomrdbox"

[[tool.importlinter.contracts]]
name = "Layered architecture"
type = "layers"
layers = [
    "boomrdbox.cli",
    "boomrdbox._config | boomrdbox.recorder | boomrdbox.player | boomrdbox.tools",
    "boomrdbox.io",
    "boomrdbox.models",
]

[[tool.importlinter.contracts]]
name = "Domain isolation"
type = "forbidden"
source_modules = ["boomrdbox.models"]
forbidden_modules = [
    "boomrdbox.io",
    "boomrdbox.recorder",
    "boomrdbox.player",
    "boomrdbox.tools",
    "boomrdbox.cli",
    "boomrdbox._config",
]
